+++ plugin-gradle/src/main/java/dev/equo/ide/gradle/EquoIdeTask.java
@@ -42,6 +42,9 @@ public abstract class EquoIdeTask extends DefaultTask {
 	@Internal
 	public abstract Property<File> getWorkspaceDir();
 
+	@Internal
+	public abstract Property<Boolean> getUseAtomos();
+
 	private boolean initOnly = false;
 
 	@Option(
@@ -49,6 +52,15 @@ public abstract class EquoIdeTask extends DefaultTask {
 			description = "Initializes the runtime to check for errors then exits.")
 	void setInitOnly(boolean initOnly) {
 		this.initOnly = initOnly;
+	}
+
+	private boolean dontUseAtomosOverride = false;
+
+	@Option(
+			option = "dont-use-atomos",
+			description = "Initializes the runtime to check for errors then exits.")
+	void dontUseAtomos(boolean dontUseAtomos) {
+		this.dontUseAtomosOverride = dontUseAtomos;
 	}
 
 	@Inject
@@ -77,12 +89,15 @@ public abstract class EquoIdeTask extends DefaultTask {
 						.collect(Collectors.toList());
 		var nestedDefs = getObjectFactory().fileCollection().from(nestedJars);
 
+		boolean useAtomos = dontUseAtomosOverride ? false : getUseAtomos().get();
 		var result =
 				NestedJars.javaExec(
 						"dev.equo.solstice.IdeMain",
 						p2AndMavenDeps.plus(nestedDefs),
 						"-installDir",
 						workspaceDir.getAbsolutePath(),
+						"-useAtomos",
+						Boolean.toString(useAtomos),
 						"-initOnly",
 						Boolean.toString(initOnly),
 						"-Dorg.slf4j.simpleLogger.defaultLogLevel=INFO");