+++ plugin-maven/src/main/java/dev/equo/ide/maven/LaunchMojo.java
@@ -46,8 +46,11 @@ import org.eclipse.aether.resolution.DependencyResult;
 public class LaunchMojo extends AbstractMojo {
 	@Component private RepositorySystem repositorySystem;
 
-	@Parameter(property = "initOnly")
-	private String initOnly;
+	@Parameter(property = "initOnly", defaultValue = "false")
+	private boolean initOnly;
+
+	@Parameter(property = "useAtomos", defaultValue = "true")
+	private boolean useAtomos;
 
 	@Parameter(property = "release")
 	private String release;
@@ -68,7 +71,6 @@ public class LaunchMojo extends AbstractMojo {
 	public void execute() throws MojoExecutionException, MojoFailureException {
 		try {
 			System.setProperty("osgi.platform", SwtPlatform.getRunning().toString());
-			boolean initOnlyTrue = "true".equals(initOnly);
 
 			List<Dependency> deps = new ArrayList<>();
 			deps.add(
@@ -89,7 +91,7 @@ public class LaunchMojo extends AbstractMojo {
 
 			var query = session.query();
 			query.platform(SwtPlatform.getRunning());
-			if (initOnlyTrue) {
+			if (initOnly) {
 				query.install("org.eclipse.swt");
 			} else {
 				JdtSetup.mavenCoordinate(query, session);
@@ -125,8 +127,10 @@ public class LaunchMojo extends AbstractMojo {
 							files,
 							"-installDir",
 							workspaceDir.getAbsolutePath(),
+							"-useAtomos",
+							Boolean.toString(useAtomos),
 							"-initOnly",
-							Boolean.toString(initOnlyTrue));
+							Boolean.toString(initOnly));
 			System.out.println(result);
 			workspaceRegistry.clean();
 		} catch (DependencyResolutionException | IOException | InterruptedException e) {