+++ plugin-gradle/CHANGELOG.md
@@ -3,6 +3,9 @@
 We adhere to the [keepachangelog](https://keepachangelog.com/en/1.0.0/) format.
 
 ## [Unreleased]
+### Added
+- Added a parameter `useAtomos` to the Solstice buildscript, default value is `true`. ([#36](https://github.com/equodev/equo-ide/pull/36))
+  - You can override the `useAtomos` buildscript parameter at the command line with `--dont-use-atomos`.
 
 ## [0.6.0] - 2022-12-16
 ### Added